Flytter feilhandteringskode.
Flytter "feilmeldinger" katalog frå v2/gui/sak til v2/gui/app. Gjev meir meining å ha det der sidan det omhandler feilmeldinger for heile appen.

Flytter ErrorBoundary komponent til v2, fikser typescript feil.
