+++ packages/sak-app/src/app/components/Home.tsx
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/react';
 import { Route, Routes } from 'react-router';
 
-import NotFoundPage from '@k9-sak-web/gui/sak/feilmeldinger/NotFoundPage.js';
+import NotFoundPage from '@k9-sak-web/gui/app/feilmeldinger/NotFoundPage.js';
 
 import AktoerIndex from '../../aktoer/AktoerIndex';
 import FagsakIndex from '../../fagsak/FagsakIndex';
