+++ packages/v2/gui/src/app/feilmeldinger/ForbiddenPage.tsx
@@ -1,16 +1,17 @@
-import { BigError } from '@k9-sak-web/gui/sak/feilmeldinger/BigError.js';
 import { Link } from 'react-router';
 
+import { BigError } from '@k9-sak-web/gui/app/feilmeldinger/BigError.js';
+
 /**
- * NotFoundPage
+ * ForbiddenPage
  *
  * Presentasjonskomponent. Denne komponenten vises når den NAV-ansatte prøver å aksessere en url som ikke finnes.
  * Det blir presentert en generell feilmelding og en lenke som tar NAV-ansatt tilbake til hovedsiden.
  */
-const NotFoundPage = () => (
-  <BigError title="Beklager, vi finner ikke siden du leter etter.">
+const ForbiddenPage = () => (
+  <BigError title="Du har ikke tilgang til å slå opp denne personen">
     <Link to="/">Gå til forsiden</Link>
   </BigError>
 );
 
-export default NotFoundPage;
+export default ForbiddenPage;
