+++ packages/v2/gui/src/app/feilmeldinger/NetworkErrorPage.stories.tsx
@@ -3,7 +3,7 @@ import { expect } from '@storybook/test';
 import NetworkErrorPage from './NetworkErrorPage.js';
 
 const meta = {
-  title: 'gui/sak/feilmeldinger',
+  title: 'gui/app/feilmeldinger',
   component: NetworkErrorPage,
 } satisfies Meta<typeof NetworkErrorPage>;
