+++ packages/v2/gui/src/app/feilmeldinger/NotFoundPage.tsx
@@ -1,17 +1,16 @@
+import { BigError } from './BigError.js';
 import { Link } from 'react-router';
 
-import { BigError } from '@k9-sak-web/gui/sak/feilmeldinger/BigError.js';
-
 /**
- * ForbiddenPage
+ * NotFoundPage
  *
  * Presentasjonskomponent. Denne komponenten vises når den NAV-ansatte prøver å aksessere en url som ikke finnes.
  * Det blir presentert en generell feilmelding og en lenke som tar NAV-ansatt tilbake til hovedsiden.
  */
-const ForbiddenPage = () => (
-  <BigError title="Du har ikke tilgang til å slå opp denne personen">
+const NotFoundPage = () => (
+  <BigError title="Beklager, vi finner ikke siden du leter etter.">
     <Link to="/">Gå til forsiden</Link>
   </BigError>
 );
 
-export default ForbiddenPage;
+export default NotFoundPage;
